requirements.txt
networkx
matplotlib
requests
google-generativeai
chromadb
pandas
google-generativeai
plotly
radon